@@ -5,7 +5,7 @@ publish_to: "none"
 # repository: https://github.com/my_org/my_repo
 
 environment:
-  sdk: '>=2.19.1 <3.0.0'
+  sdk: '>=3.10.0'
 
 dependencies:
   simple_live_core:
