@@ -4,7 +4,7 @@ description: "聚合直播核心库"
 repository: https://github.com/xiaoyaocz/dart_simple_live
 publish_to: "none"
 environment: 
-  sdk: '>=3.0.5 <4.0.0'
+  sdk: '>=3.10.0'
 
 dependencies: 
   dio: ^5.5.0+1
@@ -17,6 +17,9 @@ dependencies:
   tars_dart:
     path: ./packages/tars_dart
   fixnum: ^1.1.0
+  dart_quickjs:
+    git:
+      url: https://github.com/xiaoyaocz/dart_quickjs
 
 dev_dependencies: 
   lints: ^2.0.0
