@@ -100,23 +100,23 @@ class FollowUserPage extends GetView<FollowUserController> {
           ),
         ],
         leading: Obx(
-              () => FollowService.instance.updating.value
+          () => FollowService.instance.updating.value
               ? const IconButton(
-            onPressed: null,
-            icon: SizedBox(
-              width: 16,
-              height: 16,
-              child: CircularProgressIndicator(
-                strokeWidth: 2,
-              ),
-            ),
-          )
+                  onPressed: null,
+                  icon: SizedBox(
+                    width: 16,
+                    height: 16,
+                    child: CircularProgressIndicator(
+                      strokeWidth: 2,
+                    ),
+                  ),
+                )
               : IconButton(
-            onPressed: () {
-              controller.refreshData();
-            },
-            icon: const Icon(Icons.refresh),
-          ),
+                  onPressed: () {
+                    controller.refreshData();
+                  },
+                  icon: const Icon(Icons.refresh),
+                ),
         ),
       ),
       body: Column(
@@ -128,7 +128,7 @@ class FollowUserPage extends GetView<FollowUserController> {
               children: [
                 Expanded(
                   child: Obx(
-                        () => SingleChildScrollView(
+                    () => SingleChildScrollView(
                       scrollDirection: Axis.horizontal,
                       child: Wrap(
                           spacing: 12,
@@ -185,9 +185,9 @@ class FollowUserPage extends GetView<FollowUserController> {
       ...controller.tagList.skip(3),
     ];
     Rx<FollowUserTag> checkTag =
-    controller.tagList.indexOf(controller.filterMode.value) < 3
-        ? copiedList.first.obs
-        : controller.filterMode.value.obs;
+        controller.tagList.indexOf(controller.filterMode.value) < 3
+            ? copiedList.first.obs
+            : controller.filterMode.value.obs;
     final ScrollController scrollController = ScrollController();
     Get.dialog(
       AlertDialog(
@@ -221,7 +221,7 @@ class FollowUserPage extends GetView<FollowUserController> {
             ),
             const Divider(),
             Obx(
-                  () {
+              () {
                 int selectedIndex = copiedList.indexOf(checkTag.value);
                 WidgetsBinding.instance.addPostFrameCallback((_) {
                   if (selectedIndex >= 0) {
@@ -235,28 +235,30 @@ class FollowUserPage extends GetView<FollowUserController> {
                 return SizedBox(
                   height: 300,
                   width: 300,
-                  child: ListView.builder(
-                    controller: scrollController,
-                    itemCount: copiedList.length,
-                    itemBuilder: (context, index) {
-                      var tagItem = copiedList[index];
-                      return Container(
-                        decoration: BoxDecoration(
-                          border: Border(
-                            bottom: BorderSide(
-                                color: Colors.grey.shade300, width: 1.0),
-                          ),
-                        ),
-                        child: RadioListTile<FollowUserTag>(
-                          title: Text(tagItem.tag),
-                          value: tagItem,
-                          groupValue: checkTag.value,
-                          onChanged: (value) {
-                            checkTag.value = value!;
-                          },
-                        ),
-                      );
+                  child: RadioGroup(
+                    groupValue: checkTag.value,
+                    onChanged: (value) {
+                      checkTag.value = value!;
                     },
+                    child: ListView.builder(
+                      controller: scrollController,
+                      itemCount: copiedList.length,
+                      itemBuilder: (context, index) {
+                        var tagItem = copiedList[index];
+                        return Container(
+                          decoration: BoxDecoration(
+                            border: Border(
+                              bottom: BorderSide(
+                                  color: Colors.grey.shade300, width: 1.0),
+                            ),
+                          ),
+                          child: RadioListTile<FollowUserTag>(
+                            title: Text(tagItem.tag),
+                            value: tagItem,
+                          ),
+                        );
+                      },
+                    ),
                   ),
                 );
               },
@@ -286,7 +288,7 @@ class FollowUserPage extends GetView<FollowUserController> {
             // 列表内容
             Expanded(
               child: Obx(
-                    () => ReorderableListView.builder(
+                () => ReorderableListView.builder(
                   itemCount: controller.userTagList.length,
                   itemBuilder: (context, index) {
                     // 偏移
@@ -321,7 +323,7 @@ class FollowUserPage extends GetView<FollowUserController> {
 
   void editTagDialog(String title, {FollowUserTag? followUserTag}) {
     final TextEditingController tagEditController =
-    TextEditingController(text: followUserTag?.tag);
+        TextEditingController(text: followUserTag?.tag);
     bool upMode = title == "添加标签" ? true : false;
     Get.dialog(
       AlertDialog(
@@ -350,7 +352,7 @@ class FollowUserPage extends GetView<FollowUserController> {
                   contentPadding: AppStyle.edgeInsetsA12,
                   enabledBorder: OutlineInputBorder(
                     borderSide: BorderSide(
-                      color: Colors.grey.withOpacity(.2),
+                      color: Colors.grey.withAlpha(51),
                     ),
                   ),
                 ),
@@ -358,7 +360,7 @@ class FollowUserPage extends GetView<FollowUserController> {
                   upMode
                       ? controller.addTag(tagEditController.text)
                       : controller.updateTagName(
-                      followUserTag!, tagEditController.text);
+                          followUserTag!, tagEditController.text);
                   Get.back();
                 },
               ),
@@ -379,7 +381,7 @@ class FollowUserPage extends GetView<FollowUserController> {
                         upMode
                             ? controller.addTag(tagEditController.text)
                             : controller.updateTagName(
-                            followUserTag!, tagEditController.text);
+                                followUserTag!, tagEditController.text);
                         Get.back();
                       },
                       child: const Text('是'),
