@@ -1,7 +1,5 @@
 import 'package:simple_live_app/app/constant.dart';
 import 'package:simple_live_app/app/controller/app_settings_controller.dart';
-import 'package:simple_live_app/app/sign/douyin.dart';
-import 'package:simple_live_app/app/sign/douyu.dart';
 import 'package:simple_live_core/simple_live_core.dart';
 
 class Sites {
@@ -16,7 +14,7 @@ class Sites {
       id: Constant.kDouyu,
       logo: "assets/images/douyu.png",
       name: "斗鱼直播",
-      liveSite: DouyuSite()..setDouyuSignFunction(DouyuSign.getSign),
+      liveSite: DouyuSite(),
     ),
     Constant.kHuya: Site(
       id: Constant.kHuya,
@@ -28,9 +26,7 @@ class Sites {
       id: Constant.kDouyin,
       logo: "assets/images/douyin.png",
       name: "抖音直播",
-      liveSite: DouyinSite()
-        ..setAbogusUrlFunction(DouyinSign.getAbogusUrl)
-        ..setSignatureFunction(DouyinSign.getSignature),
+      liveSite: DouyinSite(),
     ),
   };
 
