@@ -282,20 +282,22 @@ class Utils {
     String title = '',
   }) async {
     var result = await Get.dialog(
-      SimpleDialog(
-        title: Text(title),
-        children: contents
-            .map(
-              (e) => RadioListTile<T>(
-                title: Text(e.toString()),
-                value: e,
-                groupValue: value,
-                onChanged: (e) {
-                  Get.back(result: e);
-                },
-              ),
-            )
-            .toList(),
+      RadioGroup(
+        groupValue: value,
+        onChanged: (e) {
+          Get.back(result: e);
+        },
+        child: SimpleDialog(
+          title: Text(title),
+          children: contents
+              .map(
+                (e) => RadioListTile<T>(
+                  title: Text(e.toString()),
+                  value: e,
+                ),
+              )
+              .toList(),
+        ),
       ),
     );
     return result;
@@ -315,12 +317,13 @@ class Utils {
         title: title ?? const Text("帮助"),
         scrollable: true,
         content: SingleChildScrollView(child: ListBody(children: content)),
-        actions: actions??[
-          TextButton(
-            onPressed: Get.back,
-            child: const Text("确定"),
-          ),
-        ],
+        actions: actions ??
+            [
+              TextButton(
+                onPressed: Get.back,
+                child: const Text("确定"),
+              ),
+            ],
       ),
     );
     return result;
@@ -347,25 +350,26 @@ class Utils {
     String title = '',
   }) async {
     var result = await Get.dialog(
-      SimpleDialog(
-        title: Text(title),
-        children: contents.keys
-            .map(
-              (e) => RadioListTile<T>(
-                title: Text((contents[e] ?? '-').tr),
-                value: e,
-                groupValue: value,
-                onChanged: (e) {
-                  Get.back(result: e);
-                },
-              ),
-            )
-            .toList(),
+      RadioGroup(
+        groupValue: value,
+        onChanged: (e) {
+          Get.back(result: e);
+        },
+        child: SimpleDialog(
+          title: Text(title),
+          children: contents.keys
+              .map(
+                (e) => RadioListTile<T>(
+                  title: Text((contents[e] ?? '-').tr),
+                  value: e,
+                ),
+              )
+              .toList(),
+        ),
       ),
     );
     return result;
   }
-
 
   static int parseVersion(String version) {
     var sp = version.split('.');
