@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
 import 'package:media_kit_video/media_kit_video.dart';
-import 'package:ns_danmaku/ns_danmaku.dart';
+import 'package:canvas_danmaku/canvas_danmaku.dart';
 import 'package:remixicon/remixicon.dart';
 import 'package:simple_live_app/app/app_style.dart';
 import 'package:simple_live_app/app/controller/app_settings_controller.dart';
@@ -647,11 +647,16 @@ Widget buildControls(
 
 Widget buildDanmuView(VideoState videoState, LiveRoomController controller) {
   var padding = MediaQuery.of(videoState.context).padding;
-  controller.danmakuView ??= DanmakuView(
+  controller.danmakuView ??= DanmakuScreen(
     key: controller.globalDanmuKey,
     createdController: controller.initDanmakuController,
     option: DanmakuOption(
-      fontSize: 16,
+      fontSize: AppSettingsController.instance.danmuSize.value,
+      area: AppSettingsController.instance.danmuArea.value,
+      duration: AppSettingsController.instance.danmuSpeed.value.toInt(),
+      opacity: AppSettingsController.instance.danmuOpacity.value,
+      //strokeWidth: AppSettingsController.instance.danmuStrokeWidth.value,
+      fontWeight: AppSettingsController.instance.danmuFontWeight.value,
     ),
   );
   return Positioned.fill(
@@ -788,72 +793,54 @@ void showPlayerSettings(LiveRoomController controller) {
     width: 320,
     useSystem: true,
     child: Obx(
-      () => ListView(
-        padding: AppStyle.edgeInsetsV12,
-        children: [
-          Padding(
-            padding: AppStyle.edgeInsetsH16,
-            child: Text(
-              "画面尺寸",
-              style: Get.textTheme.titleMedium,
+      () => RadioGroup(
+        groupValue: AppSettingsController.instance.scaleMode.value,
+        onChanged: (e) {
+          AppSettingsController.instance.setScaleMode(e ?? 0);
+          controller.updateScaleMode();
+        },
+        child: ListView(
+          padding: AppStyle.edgeInsetsV12,
+          children: [
+            Padding(
+              padding: AppStyle.edgeInsetsH16,
+              child: Text(
+                "画面尺寸",
+                style: Get.textTheme.titleMedium,
+              ),
             ),
-          ),
-          RadioListTile(
-            value: 0,
-            contentPadding: AppStyle.edgeInsetsH4,
-            title: const Text("适应"),
-            visualDensity: VisualDensity.compact,
-            groupValue: AppSettingsController.instance.scaleMode.value,
-            onChanged: (e) {
-              AppSettingsController.instance.setScaleMode(e ?? 0);
-              controller.updateScaleMode();
-            },
-          ),
-          RadioListTile(
-            value: 1,
-            contentPadding: AppStyle.edgeInsetsH4,
-            title: const Text("拉伸"),
-            visualDensity: VisualDensity.compact,
-            groupValue: AppSettingsController.instance.scaleMode.value,
-            onChanged: (e) {
-              AppSettingsController.instance.setScaleMode(e ?? 1);
-              controller.updateScaleMode();
-            },
-          ),
-          RadioListTile(
-            value: 2,
-            contentPadding: AppStyle.edgeInsetsH4,
-            title: const Text("铺满"),
-            visualDensity: VisualDensity.compact,
-            groupValue: AppSettingsController.instance.scaleMode.value,
-            onChanged: (e) {
-              AppSettingsController.instance.setScaleMode(e ?? 2);
-              controller.updateScaleMode();
-            },
-          ),
-          RadioListTile(
-            value: 3,
-            contentPadding: AppStyle.edgeInsetsH4,
-            title: const Text("16:9"),
-            visualDensity: VisualDensity.compact,
-            groupValue: AppSettingsController.instance.scaleMode.value,
-            onChanged: (e) {
-              AppSettingsController.instance.setScaleMode(e ?? 3);
-              controller.updateScaleMode();
-            },
-          ),
-          RadioListTile(
-            value: 4,
-            contentPadding: AppStyle.edgeInsetsH4,
-            title: const Text("4:3"),
-            visualDensity: VisualDensity.compact,
-            groupValue: AppSettingsController.instance.scaleMode.value,
-            onChanged: (e) {
-              AppSettingsController.instance.setScaleMode(e ?? 4);
-              controller.updateScaleMode();
-            },
-          ),
-        ],
+            const RadioListTile(
+              value: 0,
+              contentPadding: AppStyle.edgeInsetsH4,
+              title: Text("适应"),
+              visualDensity: VisualDensity.compact,
+            ),
+            const RadioListTile(
+              value: 1,
+              contentPadding: AppStyle.edgeInsetsH4,
+              title: Text("拉伸"),
+              visualDensity: VisualDensity.compact,
+            ),
+            const RadioListTile(
+              value: 2,
+              contentPadding: AppStyle.edgeInsetsH4,
+              title: Text("铺满"),
+              visualDensity: VisualDensity.compact,
+            ),
+            const RadioListTile(
+              value: 3,
+              contentPadding: AppStyle.edgeInsetsH4,
+              title: Text("16:9"),
+              visualDensity: VisualDensity.compact,
+            ),
+            const RadioListTile(
+              value: 4,
+              contentPadding: AppStyle.edgeInsetsH4,
+              title: Text("4:3"),
+              visualDensity: VisualDensity.compact,
+            ),
+          ],
+        ),
       ),
     ),
   );
