@@ -1,10 +1,13 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
 import 'package:get/get.dart';
 import 'package:simple_live_app/app/utils.dart';
 import 'package:simple_live_app/routes/route_path.dart';
 import 'package:simple_live_app/services/bilibili_account_service.dart';
+import 'package:simple_live_app/services/douyin_account_service.dart';
+import 'package:simple_live_core/simple_live_core.dart';
 
 class AccountController extends GetxController {
   void bilibiliTap() async {
@@ -55,7 +58,7 @@ class AccountController extends GetxController {
             trailing: const Icon(Icons.chevron_right),
             onTap: () {
               Get.back();
-              doCookieLogin();
+              doBiliBiliCookieLogin();
             },
           ),
         ],
@@ -63,7 +66,7 @@ class AccountController extends GetxController {
     );
   }
 
-  void doCookieLogin() async {
+  void doBiliBiliCookieLogin() async {
     var cookie = await Utils.showEditTextDialog(
       "",
       title: "请输入Cookie",
@@ -74,5 +77,92 @@ class AccountController extends GetxController {
     }
     BiliBiliAccountService.instance.setCookie(cookie);
     await BiliBiliAccountService.instance.loadUserInfo();
+  }
+
+  void douyinTap() async {
+    if (DouyinAccountService.instance.hasCookie.value) {
+      var result = await Utils.showAlertDialog("确定要清除自定义 ttwid 吗？", title: "清除配置");
+      if (result) {
+        DouyinAccountService.instance.clearCookie();
+        SmartDialog.showToast("已清除自定义 ttwid，将使用默认 ttwid");
+      }
+    } else {
+      doDouyinCookieConfig();
+    }
+  }
+
+  void doDouyinCookieConfig() {
+    // 初始化文本框时，只显示 ttwid 的值部分
+    var savedCookie = DouyinAccountService.instance.cookie;
+    var displayText = savedCookie;
+    if (savedCookie.startsWith('ttwid=')) {
+      displayText = savedCookie.substring(6); // 去掉 "ttwid="
+    }
+    var controller = TextEditingController(text: displayText);
+
+    Get.dialog(
+      AlertDialog(
+        title: const Text("配置抖音 ttwid"),
+        content: SingleChildScrollView(
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              const Text(
+                "默认已内置有效的 ttwid，可观看所有画质（包括蓝光）。\n如有需要可自定义配置。",
+                style: TextStyle(fontSize: 12, color: Colors.grey),
+              ),
+              const SizedBox(height: 12),
+              TextField(
+                controller: controller,
+                maxLines: 3,
+                decoration: const InputDecoration(
+                  hintText: "请粘贴 ttwid 值（留空则使用默认值）",
+                  border: OutlineInputBorder(),
+                ),
+              ),
+              const SizedBox(height: 8),
+              TextButton.icon(
+                onPressed: () {
+                  // 提取 ttwid 的值部分（去掉 "ttwid=" 前缀）
+                  var defaultValue = DouyinSite.kDefaultCookie;
+                  if (defaultValue.startsWith('ttwid=')) {
+                    defaultValue = defaultValue.substring(6); // 去掉 "ttwid="
+                  }
+                  controller.text = defaultValue;
+                },
+                icon: const Icon(Icons.restore),
+                label: const Text("恢复默认 ttwid"),
+              ),
+            ],
+          ),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Get.back(),
+            child: const Text("取消"),
+          ),
+          TextButton(
+            onPressed: () {
+              var input = controller.text.trim();
+              Get.back();
+              if (input.isEmpty) {
+                DouyinAccountService.instance.clearCookie();
+                SmartDialog.showToast("已清除自定义 Cookie，将使用默认 ttwid");
+              } else {
+                // 如果用户只输入了 ttwid 值，自动添加 "ttwid=" 前缀
+                var cookie = input;
+                if (!input.startsWith('ttwid=')) {
+                  cookie = 'ttwid=$input';
+                }
+                DouyinAccountService.instance.setCookie(cookie);
+                SmartDialog.showToast("ttwid 已保存");
+              }
+            },
+            child: const Text("确定"),
+          ),
+        ],
+      ),
+    );
   }
 }
