@@ -3,6 +3,7 @@ import 'package:get/get.dart';
 import 'package:simple_live_app/app/app_style.dart';
 import 'package:simple_live_app/modules/mine/account/account_controller.dart';
 import 'package:simple_live_app/services/bilibili_account_service.dart';
+import 'package:simple_live_app/services/douyin_account_service.dart';
 
 class AccountPage extends GetView<AccountController> {
   const AccountPage({Key? key}) : super(key: key);
@@ -18,7 +19,7 @@ class AccountPage extends GetView<AccountController> {
           const Padding(
             padding: AppStyle.edgeInsetsA12,
             child: Text(
-              "哔哩哔哩账号需要登录才能看高清晰度的直播，其他平台暂无此限制。",
+              "哔哩哔哩账号需要登录才能看高清晰度的直播。",
               textAlign: TextAlign.center,
             ),
           ),
@@ -59,16 +60,22 @@ class AccountPage extends GetView<AccountController> {
             enabled: false,
             trailing: const Icon(Icons.chevron_right),
           ),
-          ListTile(
-            leading: Image.asset(
-              'assets/images/douyin.png',
-              width: 36,
-              height: 36,
+          Obx(
+            () => ListTile(
+              leading: Image.asset(
+                'assets/images/douyin.png',
+                width: 36,
+                height: 36,
+              ),
+              title: const Text("抖音直播"),
+              subtitle: Text(DouyinAccountService.instance.hasCookie.value
+                  ? "已自定义（${DouyinAccountService.instance.cookie.length} 字符）"
+                  : "使用默认 ttwid"),
+              trailing: DouyinAccountService.instance.hasCookie.value
+                  ? const Icon(Icons.delete_outline)
+                  : const Icon(Icons.chevron_right),
+              onTap: controller.douyinTap,
             ),
-            title: const Text("抖音直播"),
-            subtitle: const Text("无需登录"),
-            enabled: false,
-            trailing: const Icon(Icons.chevron_right),
           ),
         ],
       ),
