@@ -12,7 +12,7 @@ class BiliBiliWebLoginController extends BaseController {
     webViewController = controller;
     webViewController!.loadUrl(
       urlRequest: URLRequest(
-        url: Uri.parse("https://passport.bilibili.com/login"),
+        url: WebUri("https://passport.bilibili.com/login"),
       ),
     );
   }
@@ -33,8 +33,8 @@ class BiliBiliWebLoginController extends BaseController {
 
   Future<bool> logined() async {
     try {
-      var cookies = await cookieManager.getCookies(
-          url: Uri.parse("https://bilibili.com"));
+      var cookies =
+          await cookieManager.getCookies(url: WebUri("https://bilibili.com"));
       if (cookies.isEmpty) {
         return false;
       }
