@@ -22,12 +22,10 @@ class BiliBiliWebLoginPage extends GetView<BiliBiliWebLoginController> {
       body: InAppWebView(
         onWebViewCreated: controller.onWebViewCreated,
         onLoadStop: controller.onLoadStop,
-        initialOptions: InAppWebViewGroupOptions(
-          crossPlatform: InAppWebViewOptions(
-            userAgent:
-                "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1 Edg/118.0.0.0",
-            useShouldOverrideUrlLoading: true,
-          ),
+        initialSettings: InAppWebViewSettings(
+          userAgent:
+              "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1 Edg/118.0.0.0",
+          useShouldOverrideUrlLoading: true,
         ),
         shouldOverrideUrlLoading: (webController, navigationAction) async {
           var uri = navigationAction.request.url;
