@@ -4,14 +4,30 @@ import 'dart:math';
 import 'package:simple_live_core/simple_live_core.dart';
 import 'package:simple_live_core/src/common/http_client.dart';
 import 'package:crypto/crypto.dart';
+import 'package:simple_live_core/src/model/tars/get_cdn_token_ex_req.dart';
+import 'package:simple_live_core/src/model/tars/get_cdn_token_ex_resp.dart';
 import 'package:simple_live_core/src/model/tars/get_cdn_token_req.dart';
 import 'package:simple_live_core/src/model/tars/get_cdn_token_resp.dart';
+import 'package:simple_live_core/src/model/tars/huya_user_id.dart';
 import 'package:tars_dart/tars/net/base_tars_http.dart';
 
 class HuyaSite implements LiveSite {
+  static const baseUrl = "https://m.huya.com/";
   final String kUserAgent =
       "Mozilla/5.0 (Linux; Android 11; Pixel 5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.91 Mobile Safari/537.36 Edg/117.0.0.0";
-  final BaseTarsHttp tupClient = BaseTarsHttp("http://wup.huya.com", "liveui");
+
+  static const String HYSDK_UA =
+      "HYSDK(Windows, 30000002)_APP(pc_exe&7060000&official)_SDK(trans&2.32.3.5646)";
+
+  static Map<String, String> requestHeaders =  {
+      'Origin': baseUrl,
+      'Referer': baseUrl,
+      'User-Agent': HYSDK_UA,
+  };
+
+  final BaseTarsHttp tupClient =
+  BaseTarsHttp("http://wup.huya.com", "liveui", headers: requestHeaders);
+
   String? playUserAgent;
   @override
   String id = "huya";
@@ -188,8 +204,7 @@ class HuyaSite implements LiveSite {
     } catch (e) {
       CoreLog.error(e);
     }
-    return playUserAgent ??
-        "HYSDK(Windows, 30000002)_APP(pc_exe&6080100&official)_SDK(trans&2.23.0.4969)";
+    return playUserAgent ?? HYSDK_UA;
   }
 
   @override
@@ -202,26 +217,95 @@ class HuyaSite implements LiveSite {
       var url = await getPlayUrl(line, quality.data["bitRate"]);
       ls.add(url);
     }
-    // from stream-rec url:https://github.com/stream-rec/stream-rec
-    var ua = await getHuYaUA();
+    // 最新UA需要额外验证，此方法暂时弃用
+    // var ua = await getHuYaUA();
     return LivePlayUrl(
       urls: ls,
-      headers: {"user-agent": ua},
+      headers: {"user-agent": HYSDK_UA},
     );
   }
 
   Future<String> getPlayUrl(HuyaLineModel line, int bitRate) async {
-    var req = GetCdnTokenReq();
-    req.cdnType = line.cdnType;
-    req.streamName = line.streamName;
-    var resp =
-        await tupClient.tupRequest("getCdnTokenInfo", req, GetCdnTokenResp());
-    var url =
-        '${line.line}/${resp.streamName}.flv?${resp.flvAntiCode}&codec=264';
+    var antiCode = await getCndTokenInfoEx(line.streamName);
+    antiCode = buildAntiCode(line.streamName, line.presenterUid, antiCode);
+    var url = '${line.line}/${line.streamName}.flv?${antiCode}&codec=264';
     if (bitRate > 0) {
       url += "&ratio=$bitRate";
     }
     return url;
+  }
+
+  // 构造 anticode, python转写
+  /// [stream] streamname [presenterUid] 用户id [antiCode] 页面anti
+  ///
+  /// return ture anticode
+  String buildAntiCode(String stream, int presenterUid, String antiCode) {
+    var mapAnti = Uri(query: antiCode).queryParametersAll;
+    if (!mapAnti.containsKey("fm")) {
+      return antiCode;
+    }
+
+    var ctype = mapAnti["ctype"]?.first ?? "huya_pc_exe";
+    var platformId = int.tryParse(mapAnti["t"]?.first ?? "0");
+
+    bool isWap = platformId == 103;
+    var clacStartTime = DateTime.now().millisecondsSinceEpoch;
+
+    CoreLog.i(
+        "using $presenterUid | ctype-{$ctype} | platformId - {$platformId} | isWap - {$isWap} | $clacStartTime");
+
+    var seqId = presenterUid + clacStartTime;
+    final secretHash =
+        md5.convert(utf8.encode('$seqId|$ctype|$platformId')).toString();
+
+    final convertUid = rotl64(presenterUid);
+    final calcUid = isWap ? presenterUid : convertUid;
+    final fm = Uri.decodeComponent(mapAnti['fm']!.first);
+    final secretPrefix = utf8.decode(base64.decode(fm)).split('_').first;
+    var wsTime = mapAnti['wsTime']!.first;
+    final secretStr =
+        '${secretPrefix}_${calcUid}_${stream}_${secretHash}_$wsTime';
+
+    final wsSecret = md5.convert(utf8.encode(secretStr)).toString();
+
+    final rnd = Random();
+    final ct =
+        ((int.parse(wsTime, radix: 16) + rnd.nextDouble()) * 1000).toInt();
+    final uuid = (((ct % 1e10) + rnd.nextDouble()) * 1e3 % 0xffffffff)
+        .toInt()
+        .toString();
+    final Map<String, dynamic> antiCodeRes = {
+      'wsSecret': wsSecret,
+      'wsTime': wsTime,
+      'seqid': seqId,
+      'ctype': ctype,
+      'ver': '1',
+      'fs': mapAnti['fs']!.first,
+      'fm': Uri.encodeComponent(mapAnti['fm']!.first),
+      't': platformId,
+    };
+    if (isWap) {
+      antiCodeRes.addAll({
+        'uid': presenterUid,
+        'uuid': uuid,
+      });
+    } else {
+      antiCodeRes['u'] = convertUid;
+    }
+
+    return antiCodeRes.entries.map((e) => '${e.key}=${e.value}').join('&');
+  }
+
+  /// return sFlvToken
+  Future<String> getCndTokenInfoEx(String stream) async {
+    var func = "getCdnTokenInfoEx";
+    var tid = HuyaUserId();
+    tid.sHuYaUA = "pc_exe&7060000&official";
+    var tReq = GetCdnTokenExReq();
+    tReq.tId = tid;
+    tReq.sStreamName = stream;
+    var resp = await tupClient.tupRequest(func, tReq, GetCdnTokenExResp());
+    return resp.sFlvToken;
   }
 
   @override
@@ -283,6 +367,7 @@ class HuyaSite implements LiveSite {
           hlsAntiCode: item["sHlsAntiCode"].toString(),
           streamName: item["sStreamName"].toString(),
           cdnType: item["sCdnType"].toString(),
+          presenterUid: roomInfo["topSid"]??0,
         ));
       }
     }
@@ -460,6 +545,14 @@ class HuyaSite implements LiveSite {
     return result["data"]["uid"].toString();
   }
 
+  int rotl64(int t) {
+    final low = t & 0xFFFFFFFF;
+    final rotatedLow =
+    ((low << 8) | (low >> 24)) & 0xFFFFFFFF;
+    final high = t & ~0xFFFFFFFF;
+    return high | rotatedLow;
+  }
+
   String getUUid() {
     var currentTime = DateTime.now().millisecondsSinceEpoch;
     var randomValue = Random().nextInt(4294967295);
@@ -612,6 +705,7 @@ class HuyaLineModel {
   final String streamName;
   final HuyaLineType lineType;
   int bitRate;
+  final int presenterUid; // topSid = subSid = presenterUid
 
   HuyaLineModel({
     required this.line,
@@ -621,6 +715,7 @@ class HuyaLineModel {
     required this.streamName,
     required this.cdnType,
     this.bitRate = 0,
+    required this.presenterUid,
   });
 
   @override
@@ -632,6 +727,7 @@ class HuyaLineModel {
       "hlsAntiCode": hlsAntiCode,
       "streamName": streamName,
       "lineType": lineType.toString(),
+      "presenterUid": presenterUid,
     });
   }
 }
