@@ -1,14 +1,13 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:canvas_danmaku/canvas_danmaku.dart';
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:simple_live_tv_app/app/controller/base_controller.dart';
 import 'package:get/get.dart';
 import 'package:media_kit/media_kit.dart';
 import 'package:media_kit_video/media_kit_video.dart';
-import 'package:ns_danmaku/ns_danmaku.dart';
 import 'package:simple_live_tv_app/app/controller/app_settings_controller.dart';
 import 'package:simple_live_tv_app/app/log.dart';
 import 'package:wakelock_plus/wakelock_plus.dart';
@@ -26,6 +25,16 @@ mixin PlayerMixin {
       //     AppSettingsController.instance.playerBufferSize.value * 1024 * 1024,
     ),
   );
+
+  /// 初始化播放器并设置 ao 参数
+  Future<void> initializePlayer() async {
+    var pp = player.platform as NativePlayer;
+
+    // media_kit 仓库更新导致的问题，临时解决办法
+    if (Platform.isAndroid) {
+      await pp.setProperty('force-seekable', 'yes');
+    }
+  }
 
   /// 视频控制器
   late final videoController = VideoController(
@@ -150,15 +159,15 @@ mixin PlayerDanmakuMixin on PlayerStateMixin {
 
   void initDanmakuController(DanmakuController e) {
     danmakuController = e;
-    danmakuController?.updateOption(
-      DanmakuOption(
-        fontSize: AppSettingsController.instance.danmuSize.value.w,
-        area: AppSettingsController.instance.danmuArea.value,
-        duration: AppSettingsController.instance.danmuSpeed.value,
-        opacity: AppSettingsController.instance.danmuOpacity.value,
-        strokeWidth: AppSettingsController.instance.danmuStrokeWidth.value.w,
-      ),
-    );
+    // danmakuController?.updateOption(
+    //   DanmakuOption(
+    //     fontSize: AppSettingsController.instance.danmuSize.value.w,
+    //     area: AppSettingsController.instance.danmuArea.value,
+    //     duration: AppSettingsController.instance.danmuSpeed.value,
+    //     opacity: AppSettingsController.instance.danmuOpacity.value,
+    //     strokeWidth: AppSettingsController.instance.danmuStrokeWidth.value.w,
+    //   ),
+    // );
   }
 
   void updateDanmuOption(DanmakuOption? option) {
@@ -170,11 +179,13 @@ mixin PlayerDanmakuMixin on PlayerStateMixin {
     danmakuController?.clear();
   }
 
-  void addDanmaku(List<DanmakuItem> items) {
+  void addDanmaku(List<DanmakuContentItem> items) {
     if (!showDanmakuState.value) {
       return;
     }
-    danmakuController?.addItems(items);
+    for (var item in items) {
+      danmakuController?.addDanmaku(item);
+    }
   }
 }
 mixin PlayerSystemMixin on PlayerMixin, PlayerStateMixin, PlayerDanmakuMixin {
