@@ -1,8 +1,8 @@
+import 'package:canvas_danmaku/canvas_danmaku.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:media_kit_video/media_kit_video.dart';
-import 'package:ns_danmaku/ns_danmaku.dart';
 import 'package:simple_live_tv_app/app/app_focus_node.dart';
 import 'package:simple_live_tv_app/app/app_style.dart';
 import 'package:simple_live_tv_app/app/controller/app_settings_controller.dart';
@@ -15,29 +15,18 @@ import 'package:simple_live_tv_app/widgets/card/anchor_card.dart';
 import 'package:simple_live_tv_app/widgets/settings_item_widget.dart';
 import 'package:simple_live_tv_app/widgets/status/app_empty_widget.dart';
 
-Widget playerControls(
-  VideoState videoState,
-  LiveRoomController controller,
-) {
-  return buildControls(
-    videoState,
-    controller,
-  );
+Widget playerControls(VideoState videoState, LiveRoomController controller) {
+  return buildControls(videoState, controller);
 }
 
-Widget buildControls(
-  VideoState videoState,
-  LiveRoomController controller,
-) {
+Widget buildControls(VideoState videoState, LiveRoomController controller) {
   return Stack(
     children: [
       Container(),
       buildDanmuView(videoState, controller),
       // 点击播放器打开设置
       Positioned.fill(
-        child: GestureDetector(
-          onTap: () => showPlayerSettings(controller),
-        ),
+        child: GestureDetector(onTap: () => showPlayerSettings(controller)),
       ),
       Center(
         child: // 中间
@@ -79,10 +68,7 @@ Widget buildControls(
               gradient: LinearGradient(
                 begin: Alignment.bottomCenter,
                 end: Alignment.topCenter,
-                colors: [
-                  Colors.transparent,
-                  Colors.black87,
-                ],
+                colors: [Colors.transparent, Colors.black87],
               ),
             ),
             child: Row(
@@ -102,9 +88,7 @@ Widget buildControls(
                         "${controller.detail.value?.userName} · ${controller.followed.value ? "已关注" : "未关注"}",
                         maxLines: 1,
                         overflow: TextOverflow.ellipsis,
-                        style: AppStyle.textStyleWhite.copyWith(
-                          fontSize: 24.w,
-                        ),
+                        style: AppStyle.textStyleWhite.copyWith(fontSize: 24.w),
                       ),
                     ],
                   ),
@@ -135,10 +119,7 @@ Widget buildControls(
               gradient: LinearGradient(
                 begin: Alignment.topCenter,
                 end: Alignment.bottomCenter,
-                colors: [
-                  Colors.transparent,
-                  Colors.black87,
-                ],
+                colors: [Colors.transparent, Colors.black87],
               ),
             ),
             padding: EdgeInsets.only(
@@ -153,14 +134,18 @@ Widget buildControls(
                   children: [
                     //清晰度
                     Obx(
-                      () => Text("清晰度: ${controller.currentQualityInfo.value}",
-                          style: AppStyle.textStyleWhite),
+                      () => Text(
+                        "清晰度: ${controller.currentQualityInfo.value}",
+                        style: AppStyle.textStyleWhite,
+                      ),
                     ),
                     //线路
                     AppStyle.hGap32,
                     Obx(
-                      () => Text("线路：${controller.currentLineInfo.value}",
-                          style: AppStyle.textStyleWhite),
+                      () => Text(
+                        "线路：${controller.currentLineInfo.value}",
+                        style: AppStyle.textStyleWhite,
+                      ),
                     ),
                     //弹幕开关
                     AppStyle.hGap32,
@@ -190,10 +175,7 @@ Widget buildControls(
                       size: 40.w,
                     ),
                     AppStyle.hGap16,
-                    Text(
-                      "上一频道",
-                      style: AppStyle.textStyleWhite,
-                    ),
+                    Text("上一频道", style: AppStyle.textStyleWhite),
                     AppStyle.hGap32,
                     Icon(
                       Icons.arrow_circle_down_rounded,
@@ -201,10 +183,7 @@ Widget buildControls(
                       size: 40.w,
                     ),
                     AppStyle.hGap16,
-                    Text(
-                      "下一频道",
-                      style: AppStyle.textStyleWhite,
-                    ),
+                    Text("下一频道", style: AppStyle.textStyleWhite),
                     AppStyle.hGap32,
                     Icon(
                       Icons.arrow_circle_left_outlined,
@@ -212,10 +191,7 @@ Widget buildControls(
                       size: 40.w,
                     ),
                     AppStyle.hGap16,
-                    Text(
-                      "关注列表",
-                      style: AppStyle.textStyleWhite,
-                    ),
+                    Text("关注列表", style: AppStyle.textStyleWhite),
                     AppStyle.hGap32,
                     Icon(
                       Icons.arrow_circle_right_outlined,
@@ -223,16 +199,9 @@ Widget buildControls(
                       size: 40.w,
                     ),
                     AppStyle.hGap16,
-                    Icon(
-                      Icons.menu,
-                      color: Colors.white,
-                      size: 44.w,
-                    ),
+                    Icon(Icons.menu, color: Colors.white, size: 44.w),
                     AppStyle.hGap16,
-                    Text(
-                      "设置",
-                      style: AppStyle.textStyleWhite,
-                    ),
+                    Text("设置", style: AppStyle.textStyleWhite),
                   ],
                 ),
               ],
@@ -246,11 +215,14 @@ Widget buildControls(
 
 Widget buildDanmuView(VideoState videoState, LiveRoomController controller) {
   var padding = MediaQuery.of(videoState.context).padding;
-  controller.danmakuView ??= DanmakuView(
+  controller.danmakuView ??= DanmakuScreen(
     key: controller.globalDanmuKey,
     createdController: controller.initDanmakuController,
     option: DanmakuOption(
-      fontSize: 40.w,
+      fontSize: AppSettingsController.instance.danmuSize.value.w,
+      area: AppSettingsController.instance.danmuArea.value,
+      duration: AppSettingsController.instance.danmuSpeed.value.toInt(),
+      opacity: AppSettingsController.instance.danmuOpacity.value,
     ),
   );
   return Positioned.fill(
@@ -362,7 +334,6 @@ void showPlayerSettings(LiveRoomController controller) {
   var danmakuSpeedFoucsNode = AppFocusNode();
   var danmakuAreaFoucsNode = AppFocusNode();
   var danmakuOpacityFoucsNode = AppFocusNode();
-  var danmakuStorkeFoucsNode = AppFocusNode();
   Utils.showSystemRightDialog(
     width: 800.w,
     child: Column(
@@ -402,10 +373,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   foucsNode: followFocusNode,
                   autofocus: followFocusNode.isFoucsed.value,
                   title: "关注用户",
-                  items: const {
-                    false: "否",
-                    true: "是",
-                  },
+                  items: const {false: "否", true: "是"},
                   value: controller.followed.value,
                   onChanged: (e) {
                     if (e) {
@@ -417,10 +385,7 @@ void showPlayerSettings(LiveRoomController controller) {
                 ),
               ),
 
-              Divider(
-                color: Colors.grey.withAlpha(50),
-                height: 36.w,
-              ),
+              Divider(color: Colors.grey.withAlpha(50), height: 36.w),
               AppStyle.vGap24,
               Padding(
                 padding: AppStyle.edgeInsetsH20,
@@ -441,12 +406,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   autofocus: qualityFoucsNode.isFoucsed.value,
                   items: controller.qualites
                       .asMap()
-                      .map(
-                        (i, e) => MapEntry(
-                          i,
-                          e.quality,
-                        ),
-                      )
+                      .map((i, e) => MapEntry(i, e.quality))
                       .cast<int, String>(),
                   value: controller.currentQuality,
                   onChanged: (e) {
@@ -466,12 +426,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   autofocus: lineFoucsNode.isFoucsed.value,
                   items: controller.playUrls
                       .asMap()
-                      .map(
-                        (i, e) => MapEntry(
-                          i,
-                          "线路${i + 1}",
-                        ),
-                      )
+                      .map((i, e) => MapEntry(i, "线路${i + 1}"))
                       .cast<int, String>(),
                   value: controller.currentLineIndex,
                   onChanged: (e) {
@@ -480,10 +435,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   },
                 ),
               ),
-              Divider(
-                color: Colors.grey.withAlpha(50),
-                height: 36.w,
-              ),
+              Divider(color: Colors.grey.withAlpha(50), height: 36.w),
               Padding(
                 padding: AppStyle.edgeInsetsH20,
                 child: Text(
@@ -499,13 +451,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   foucsNode: scaleFoucsNode,
                   autofocus: scaleFoucsNode.isFoucsed.value,
                   title: "画面比例",
-                  items: const {
-                    0: "适应",
-                    1: "拉伸",
-                    2: "铺满",
-                    3: "16:9",
-                    4: "4:3",
-                  },
+                  items: const {0: "适应", 1: "拉伸", 2: "铺满", 3: "16:9", 4: "4:3"},
                   value: AppSettingsController.instance.scaleMode.value,
                   onChanged: (e) {
                     AppSettingsController.instance.setScaleMode(e);
@@ -513,10 +459,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   },
                 ),
               ),
-              Divider(
-                color: Colors.grey.withAlpha(50),
-                height: 36.w,
-              ),
+              Divider(color: Colors.grey.withAlpha(50), height: 36.w),
               Padding(
                 padding: AppStyle.edgeInsetsH20,
                 child: Text(
@@ -532,10 +475,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   foucsNode: danmakuFoucsNode,
                   autofocus: danmakuFoucsNode.isFoucsed.value,
                   title: "弹幕开关",
-                  items: const {
-                    0: "关",
-                    1: "开",
-                  },
+                  items: const {0: "关", 1: "开"},
                   value: controller.showDanmakuState.value ? 1 : 0,
                   onChanged: (e) {
                     controller.showDanmakuState.value = e == 1;
@@ -589,7 +529,7 @@ void showPlayerSettings(LiveRoomController controller) {
                     AppSettingsController.instance.setDanmuSpeed(e);
                     controller.updateDanmuOption(
                       controller.danmakuController?.option.copyWith(
-                        duration: e as double,
+                        duration: (e as double).toInt(),
                       ),
                     );
                   },
@@ -602,12 +542,7 @@ void showPlayerSettings(LiveRoomController controller) {
                   foucsNode: danmakuAreaFoucsNode,
                   autofocus: danmakuAreaFoucsNode.isFoucsed.value,
                   title: "显示区域",
-                  items: {
-                    0.25: "1/4",
-                    0.5: "1/2",
-                    0.75: "3/4",
-                    1.0: "全屏",
-                  },
+                  items: {0.25: "1/4", 0.5: "1/2", 0.75: "3/4", 1.0: "全屏"},
                   value: AppSettingsController.instance.danmuArea.value,
                   onChanged: (e) {
                     AppSettingsController.instance.setDanmuArea(e);
@@ -648,47 +583,25 @@ void showPlayerSettings(LiveRoomController controller) {
                   },
                 ),
               ),
-              AppStyle.vGap24,
-              Obx(
-                () => SettingsItemWidget(
-                  foucsNode: danmakuStorkeFoucsNode,
-                  autofocus: danmakuStorkeFoucsNode.isFoucsed.value,
-                  title: "描边宽度",
-                  items: {
-                    2.0: "2",
-                    4.0: "4",
-                    6.0: "6",
-                    8.0: "8",
-                    10.0: "10",
-                    12.0: "12",
-                    14.0: "14",
-                    16.0: "16",
-                  },
-                  value: AppSettingsController.instance.danmuStrokeWidth.value,
-                  onChanged: (e) {
-                    AppSettingsController.instance.setDanmuStrokeWidth(e);
-                    controller.updateDanmuOption(
-                      controller.danmakuController?.option.copyWith(
-                        strokeWidth: (e as double).w,
-                      ),
-                    );
-                  },
-                ),
-              ),
             ],
           ),
         ),
       ],
     ),
-  );
+  ).then((value) {
+    // 还原焦点
+    controller.focusNode.requestFocus();
+  });
 }
 
 void showFollowUser(LiveRoomController controller) {
   var currentIndex = 0;
   if (controller.followed.value) {
-    currentIndex = FollowUserService.instance.livingList.indexWhere((e) =>
-        e.id ==
-        "${controller.rxSite.value.id}_${controller.detail.value?.roomId}");
+    currentIndex = FollowUserService.instance.livingList.indexWhere(
+      (e) =>
+          e.id ==
+          "${controller.rxSite.value.id}_${controller.detail.value?.roomId}",
+    );
     if (currentIndex == -1) {
       currentIndex = 0;
     }
@@ -733,8 +646,10 @@ void showFollowUser(LiveRoomController controller) {
                 () => ListView.separated(
                   itemCount: FollowUserService.instance.livingList.length,
                   separatorBuilder: (context, index) => AppStyle.vGap32,
-                  padding:
-                      AppStyle.edgeInsetsA40.copyWith(left: 48.w, right: 48.w),
+                  padding: AppStyle.edgeInsetsA40.copyWith(
+                    left: 48.w,
+                    right: 48.w,
+                  ),
                   itemBuilder: (_, i) {
                     var item = FollowUserService.instance.livingList[i];
                     var site = Sites.allSites[item.siteId]!;
@@ -756,9 +671,7 @@ void showFollowUser(LiveRoomController controller) {
               Obx(
                 () => Visibility(
                   visible: FollowUserService.instance.list.isEmpty,
-                  child: const AppEmptyWidget(
-                    text: "关注列表为空，快去关注一些主播吧",
-                  ),
+                  child: const AppEmptyWidget(text: "关注列表为空，快去关注一些主播吧"),
                 ),
               ),
             ],
@@ -766,5 +679,8 @@ void showFollowUser(LiveRoomController controller) {
         ),
       ],
     ),
-  );
+  ).then((value) {
+    // 还原焦点
+    controller.focusNode.requestFocus();
+  });
 }
